SD_backend/requirements.txt
Django==4.1.7
django-filter==23.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
django-cors-headers==3.14.0
mysqlclient==2.1.1
PyJWT==2.6.0
uwsgi==2.0.21